+++ pyout/tabular_dummy.py
@@ -45,10 +45,12 @@ class Tabular(interface.Writer):
     """
 
     def __init__(self, columns=None, style=None, stream=None,
-                 interactive=None, mode=None, max_workers=None):
+                 interactive=None, mode=None, continue_on_failure=True,
+                 max_workers=None):
         super(Tabular, self).__init__(
             columns, style, stream=stream,
             interactive=interactive, mode=mode,
+            continue_on_failure=continue_on_failure,
             max_workers=max_workers)
         streamer = NoUpdateTerminalStream(
             stream=stream, interactive=interactive)